Move the Windows implementation to the `internal` package
