+++ core/windows/src/internal/TzdbInRegistry.kt
@@ -3,8 +3,9 @@
  * Use of this source code is governed by the Apache 2.0 License that can be found in the LICENSE.txt file.
  */
 @file:OptIn(kotlinx.cinterop.ExperimentalForeignApi::class)
-package kotlinx.datetime
+package kotlinx.datetime.internal
 
+import kotlinx.datetime.*
 import kotlinx.cinterop.*
 import kotlinx.datetime.internal.*
 import platform.windows.*
